Stepped back to 1.15.2
We are not going to update to 1.16 until the recommended version of forge is released; this is because crucial aspects of forge that this mod uses are being revamped (world gen; biomes etc.).

Found some bugs in the previous version of this mod so WHOOPS here we go again back to the dev board.