Fixed minor bug with rice crop
Previously when using this mod with the Straw Golems or Harvest Sprites mod, when they would attempt to replant the rice crop they would remove the water block and plant rice in air (which shouldn't be possible and glithes out when recieving an update).

This fixes that issue: When planting rice, its default waterlogged state is now true, instead of being false and then immediately updating to true - which was inefficient.